more bustage fixes
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@925 0039d316-1c4b-4281-b951-d872f2087c98

+++ base/pickle_unittest.cc
@@ -189,14 +189,15 @@ TEST(PickleTest, Resize) {
   // note that any data will have a 4-byte header indicating the size
   const size_t payload_size_after_header = unit - sizeof(uint32);
   Pickle pickle;
-  pickle.WriteData(data_ptr, payload_size_after_header - sizeof(uint32));
+  pickle.WriteData(data_ptr,
+      static_cast<int>(payload_size_after_header - sizeof(uint32)));
   size_t cur_payload = payload_size_after_header;
 
   EXPECT_EQ(pickle.capacity(), unit);
   EXPECT_EQ(pickle.payload_size(), payload_size_after_header);
 
   // fill out a full page (noting data header)
-  pickle.WriteData(data_ptr, unit - sizeof(uint32));
+  pickle.WriteData(data_ptr, static_cast<int>(unit - sizeof(uint32)));
   cur_payload += unit;
   EXPECT_EQ(unit * 2, pickle.capacity());
   EXPECT_EQ(cur_payload, pickle.payload_size());
